refactor: 抽离歌词搜索服务
